Continued to add more methods to the various classes. Most have been getter and setter methods, although there have been a few based on logic.
Also added a few more attributes and removed a couple. Cats, for example, do not need a size category because that will go into the characteristics dictionary.

I am continuing to think about this program when I'm not working on it. It continues to expand, and I'm happy to find that I have a greater understanding of what types of elements I need in the game than I originally feared.